Rename mock data to `searchResultsData`
