+++ src/components/SearchResults.js
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import styled from 'styled-components';
 
 import ListingCard from './ListingCard';
-import { get } from '../mockBackend/fetch';
+import { getSearchResults as get } from '../mockBackend/fetch';
 
 const Container = styled.div`
   background-color: #E5E5E5;
