+++ src/mockBackend/fetch.js
@@ -6,14 +6,16 @@
 // In a real app, this would grab data from the internet, but
 // this module just waits a little bit before responding.
 
-import DATA from "./data";
+import searchResultsData from "./searchResultsData";
 
-export function get() {
+const getSearchResults = () => {
   const delay = Math.floor(Math.random() * 1000);
   return new Promise((resolve, reject) => {
     setTimeout(() => {
-      const response = { status: 200, data: DATA };
+      const response = { status: 200, data: searchResultsData };
       resolve(response);
     }, delay);
   });
-}
+};
+
+export { getSearchResults };
